# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "安然画室 · 课程预约小程序 API",
    "version": "1.0.0",
    "description": "线下课程介绍+预约微信小程序后端接口文档。包含用户端公开接口、用户端需登录接口、管理后台接口三部分。\n\n**认证方式**: JWT Bearer Token（Header: `Authorization: Bearer <token>`）\n\n**管理后台**: 用户名密码登录，使用 nickmame 作为用户名，默认密码为环境变量 `ADMIN_PASSWORD`（默认 `admin123`）\n\n**响应格式**: 统一为 `{ code: 0, message: \"ok\", data: {} }`\n\n**开发测试账号**: openid=dev_test_openid，无需真实微信code即可登录"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.anran-studio.com/api",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，登录后获取。用户端使用微信code登录，管理端使用用户名密码登录。"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "data": {
            "type": "null",
            "description": "始终为 null"
          }
        }
      },
...
      "CourseSchedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "course_type_id": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "上课日期"
          },
          "start_time": {
            "type": "string",
            "description": "开始时间（HH:mm:ss）"
          },
          "end_time": {
            "type": "string",
            "description": "结束时间（HH:mm:ss）"
          },
          "max_count": {
            "type": "integer",
            "description": "最大容量"
          },
          "booked_count": {
            "type": "integer",
            "description": "已预约人数"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "full"
            ],
            "description": "时段状态"
          },
          "check_in_code": {
            "type": "string",
            "nullable": true,
            "description": "核销码（预留）"
          },
          "checked_in_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CourseType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "course_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "班型名称（如：启蒙班/进阶班）"
          },
          "capacity": {
            "type": "integer",
            "description": "班级容量"
          },
          "price": {
            "type": "number",
            "description": "价格（预留）"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "学员姓名"
          },
          "age": {
            "type": "integer",
            "description": "年龄"
          },
          "phone": {
            "type": "string",
            "description": "联系电话"
          },
          "notes": {
            "type": "string",
            "description": "备注"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录或 Token 无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "noToken": {
                "value": {
                  "code": 401,
                  "message": "请先登录",
                  "data": null
                }
              },
              "expired": {
                "value": {
                  "code": 401,
                  "message": "登录已过期，请重新登录",
                  "data": null
                }
              },
              "invalid": {
                "value": {
                  "code": 401,
                  "message": "无效的登录凭证",
                  "data": null
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "无管理员权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": 403,
              "message": "无管理员权限",
              "data": null
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/schedules/{id}": {
      "get": {
        "tags": [
          "时段"
        ],
        "summary": "时段详情",
        "description": "获取课程时段详情，包含剩余名额和关联的班型/课程信息。",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "时段ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "时段详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/CourseSchedule"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "courseType": {
                              "allOf": [
                                {
                                  "$ref": "#/components/schemas/CourseType"
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "course": {
                                      "type": "object",
                                      "properties": {
                                        "id": {
                                          "type": "integer"
                                        },
                                        "title": {
                                          "type": "string"
                                        }
                                      }
                                    }
                                  }
                                }
                              ]
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "时段不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "时段不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
    "/students": {
      "get": {
        "tags": [
          "学员管理"
        ],
        "summary": "我的学员列表",
        "description": "获取当前用户添加的所有学员档案。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "学员列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Student"
                      },
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "code": 0,
                  "message": "ok",
                  "data": [
                    {
                      "id": 1,
                      "user_id": 1,
                      "name": "朵朵",
                      "age": 5,
                      "phone": "13800138000",
                      "notes": "",
                      "created_at": "2026-05-01T00:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "学员管理"
        ],
        "summary": "添加学员",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "学员姓名",
                    "example": "朵朵"
                  },
                  "age": {
                    "type": "integer",
                    "description": "年龄",
                    "default": 0,
                    "example": 5
                  },
                  "phone": {
                    "type": "string",
                    "description": "联系电话",
                    "default": "",
                    "example": "13800138000"
                  },
                  "notes": {
                    "type": "string",
                    "description": "备注信息",
                    "default": "",
                    "example": "对美术特别感兴趣"
                  }
                }
              },
              "example": {
                "name": "朵朵",
                "age": 5,
                "phone": "13800138000",
                "notes": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "添加成功"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Student"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "姓名不能为空",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "学员姓名不能为空",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/students/{id}": {
      "put": {
        "tags": [
          "学员管理"
        ],
        "summary": "编辑学员信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "学员ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "学员姓名"
                  },
                  "age": {
                    "type": "integer",
                    "description": "年龄"
                  },
                  "phone": {
                    "type": "string",
                    "description": "联系电话"
                  },
                  "notes": {
                    "type": "string",
                    "description": "备注"
                  }
                }
              },
              "example": {
                "name": "朵朵",
                "age": 6,
                "phone": "13800138000",
                "notes": "已升班"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "更新成功"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Student"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "学员不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "学员不存在",
                  "data": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "学员管理"
        ],
        "summary": "删除学员",
        "description": "删除当前用户名下的学员档案。注意：当前后端未限制有关联预约记录的学员不可删除，实际使用中请谨慎。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "学员ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "删除成功",
                  "data": null
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "nullable": true,
                      "description": "操作类接口成功时返回 null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "学员不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "学员不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/courses/schedules/{id}": {
      "put": {
        "tags": [
          "管理端-课程"
        ],
        "summary": "编辑时段",
        "description": "修改时段信息。修改max_count时不能低于已预约人数。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "时段ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "start_time": {
                    "type": "string"
                  },
                  "end_time": {
                    "type": "string"
                  },
                  "max_count": {
                    "type": "integer",
                    "description": "最大容量（不能低于已预约数）"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "full"
                    ]
                  }
                }
              },
              "example": {
                "date": "2026-05-17",
                "start_time": "10:00:00",
                "end_time": "12:00:00"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "时段更新成功",
                  "data": {
                    "id": 1,
                    "date": "2026-05-17",
                    "start_time": "10:00:00",
                    "end_time": "12:00:00"
                  }
                }
              }
            }
          },
          "400": {
            "description": "名额不能低于已预约人数",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "名额不能低于已预约人数",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "时段不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "时段不存在",
                  "data": null
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理端-课程"
        ],
        "summary": "删除时段",
        "description": "删除时段。已有预约时禁止删除。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "时段ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "时段已删除",
                  "data": null
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "nullable": true,
                      "description": "操作类接口成功时返回 null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "该时段已有预约",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "该时段已有预约，无法删除",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}